[@dhealthdapps/backend] refactor: extract activities from processor scope to users scope
