+++ runtime/backend/src/statistics/schedulers/LeaderboardAggregation/WeeklyScoreAggegation.ts
@@ -28,11 +28,8 @@ import {
   AssetModel,
 } from "../../../discovery/models/AssetSchema";
 
-// processor scope
-import {
-  Activity,
-  ActivityModel,
-} from "../../../processor/models/ActivitySchema";
+// users scope
+import { Activity, ActivityModel } from "../../../users/models/ActivitySchema";
 
 // statistics scope
 import { LeaderboardAggregation } from "./LeaderboardAggregation";
