+++ runtime/backend/src/users/modules/ActivitiesModule.ts
@@ -19,7 +19,7 @@ import { ActivitiesController } from "../routes/ActivitiesController";
 import { ActivitiesService } from "../services/ActivitiesService";
 
 /**
- * @label PROCESSOR
+ * @label USER
  * @class ActivitiesModule
  * @description The main definition for the Activities module.
  *
