+++ runtime/backend/tests/unit/common/ScopeFactory.spec.ts
@@ -110,11 +110,6 @@ jest.mock("../../../src/processor/modules/OperationsModule", () => {
   return { OperationsModule: OperationsModuleMock };
 });
 
-const ActivitiesModuleMock: any = jest.fn();
-jest.mock("../../../src/processor/modules/ActivitiesModule", () => {
-  return { ActivitiesModule: ActivitiesModuleMock };
-});
-
 const WebHooksModuleMock: any = jest.fn();
 jest.mock("../../../src/oauth/modules/WebHooksModule", () => {
   return { WebHooksModule: WebHooksModuleMock };
@@ -152,6 +147,17 @@ jest.mock("../../../src/notifier/NotifierModule", () => {
 const OAuthModuleMock: any = jest.fn();
 jest.mock("../../../src/oauth/OAuthModule", () => {
   return { OAuthModule: OAuthModuleMock };
+});
+
+// users scope
+const ActivitiesModuleMock: any = jest.fn();
+jest.mock("../../../src/users/modules/ActivitiesModule", () => {
+  return { ActivitiesModule: ActivitiesModuleMock };
+});
+
+const UsersModuleMock: any = jest.fn();
+jest.mock("../../../src/users/UsersModule", () => {
+  return { UsersModule: UsersModuleMock };
 });
 
 // schedulers
@@ -327,7 +333,7 @@ describe("common/ScopeFactory", () => {
       const configDto: DappConfig = {
         dappName: "Fake dApp",
         dappPublicKey: "FakePublicKeyOfAdApp",
-        scopes: ["discovery", "payout", "processor", "statistics", "notifier", "oauth"],
+        scopes: ["discovery", "payout", "processor", "statistics", "notifier", "oauth", "users"],
         database: {
           host: "fake",
           port: "1",
@@ -360,6 +366,7 @@ describe("common/ScopeFactory", () => {
         StatisticsModuleMock,
         NotifierModuleMock,
         OAuthModuleMock,
+        UsersModuleMock,
       ]);
     });
