+++ runtime/backend/tests/unit/oauth/services/WebHooksService.spec.ts
@@ -18,8 +18,8 @@ import { QueryService } from "../../../../src/common/services/QueryService";
 import { CipherService } from "../../../../src/common/services/CipherService";
 import { OAuthService } from "../../../../src/oauth/services/OAuthService";
 import { WebHooksService } from "../../../../src/oauth/services/WebHooksService";
-import { ActivitiesService } from "../../../../src/processor/services/ActivitiesService";
-import { ActivityDocument, ActivityModel, ActivityQuery } from "../../../../src/processor/models/ActivitySchema";
+import { ActivitiesService } from "../../../../src/users/services/ActivitiesService";
+import { ActivityDocument, ActivityModel, ActivityQuery } from "../../../../src/users/models/ActivitySchema";
 import { StravaWebHookEventRequest } from "../../../../src/oauth/drivers/strava/StravaWebHookEventRequest";
 import { OnActivityCreated } from "../../../../src/oauth/events/OnActivityCreated";
 import { LogService } from "../../../../src/common/services/LogService";
